initial commit
Tested software build environment is Arduino 1.8.2 on Windows 10 x64.
Most ESP8266 or ESP8285 modules  will work as long as they bring out the GPIO pins that are used in this project.  I used an ESP8285 ESP-M2 module because it's literally thumbnail sized and I wanted everything to fit in a small case.  If using an ESP8266 module, select "Generic ESP8266 module" in Arduino.  If you use a development board such as the Wemos D1 or Nodemcu, bear in mind that the onboard USB-UART chip will draw power even when the ESP8266 and MPU9250 are in sleep mode, so the extra current drain will be a problem for the auto power off feature, as well as for normal active mode.  

The circuit schematic does not show the microUSB->Lipoly charger module that I used. You can use any USB->Li-ion charger module, but ensure the charge current is appropriate for the battery you use. E.g. for a 500mAH battery, ensure the maximum programming current is set to less than 250mA.  This is usually done by changing a resisto